+++ packages/prettier-config/index.js
@@ -22,7 +22,7 @@ module.exports = {
     {
       files: '*.{ts,tsx}',
       options: {
-        parser: 'babel-ts',
+        parser: 'typescript',
       },
     },
   ],